requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssmin==0.2.0
cssprefixer==1.4.1
cssutils==1.0.2
Django==2.2.7
docutils==0.15.2
Flask==1.1.1
Flask-Assets==0.12
gevent==1.4.0
gevent-socketio-hartwork==0.3.6.post201803091952
gevent-websocket==0.10.1
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lockfile==0.12.2
Logbook==1.5.3
MarkupSafe==1.1.1
passlib==1.7.2
pyScss==1.3.5
python-daemon==2.2.4
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.16.0
